# app/docs/swagger.yaml
definitions:
  app_api_handlers_controllers.DespoitRequest:
    properties:
      amount:
        type: string
      symbol:
        type: string
      user_id:
        type: string
    type: object
  app_api_handlers_controllers.TransferRequest:
    properties:
      amount:
        type: string
      from:
        type: string
      symbol:
        type: string
      to:
        type: string
    type: object
  app_api_handlers_controllers.WithdrawRequest:
    properties:
      amount:
        type: string
      symbol:
        type: string
      user_id:
        type: string
    type: object
  github_com_yzimhao_trading_engine_v2_internal_persistence_gorm_entities.Asset:
    properties:
      UUID:
        type: string
      avail_balance:
        type: string
      created_at:
        type: string
      freeze_balance:
        type: string
      symbol:
        type: string
      total_balance:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/asset/{symbol}:
    get:
      consumes:
      - application/json
      description: get an asset balance
      operationId: v1.asset.query
      parameters:
      - description: symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yzimhao_trading_engine_v2_internal_persistence_gorm_entities.Asset'
      summary: get wallet asset
      tags:
      - asset
  /api/v1/asset/{symbol}/history:
    get:
      consumes:
      - application/json
      description: get an asset history
      operationId: v1.asset.history
      parameters:
      - description: symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_yzimhao_trading_engine_v2_internal_persistence_gorm_entities.Asset'
            type: array
      summary: get asset history
      tags:
      - asset
  /api/v1/asset/despoit:
    post:
      consumes:
      - application/json
      description: despoit an asset
      operationId: v1.asset.despoit
      parameters:
      - description: despoit request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_api_handlers_controllers.DespoitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: asset despoit
      tags:
      - asset
  /api/v1/asset/transfer/{symbol}:
    post:
      consumes:
      - application/json
      description: transfer an asset
      operationId: v1.asset.transfer
      parameters:
      - description: symbol
        in: path
        name: symbol
        required: true
        type: string
      - description: transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_api_handlers_controllers.TransferRequest'
      produces:
      - application/json
      responses: {}
      summary: asset transfer
      tags:
      - asset
  /api/v1/asset/withdraw:
    post:
      consumes:
      - application/json
      description: withdraw an asset
      operationId: v1.asset.withdraw
      parameters:
      - description: withdraw request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_api_handlers_controllers.WithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: asset withdraw
      tags:
      - asset
  /api/v1/base/exchange_info:
    get:
      consumes:
      - application/json
      description: get exchange info
      operationId: v1.base.exchange_info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: exchange info
      tags:
      - base
  /api/v1/market/depth:
    get:
      consumes:
      - application/json
      description: get depth
      operationId: v1.market.depth
      parameters:
      - description: symbol
        in: query
        name: symbol
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: depth
      tags:
      - market
  /api/v1/market/klines:
    get:
      consumes:
      - application/json
      description: 获取K线数据
      operationId: v1.market.klines
      parameters:
      - description: symbol
        in: query
        name: symbol
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: klines
      tags:
      - market
  /api/v1/market/trades:
    get:
      consumes:
      - application/json
      description: 获取近期成交记录
      operationId: v1.market.trades
      parameters:
      - description: symbol
        in: query
        name: symbol
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: trades
      tags:
      - market
  /api/v1/ping:
    get:
      consumes:
      - application/json
      description: test if the server is running
      operationId: v1.ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping
      tags:
      - base
swagger: "2.0"